fix(ci): adds  new secrets for build in ci
