+++ .github/workflows/ci.yml
@@ -38,7 +38,7 @@ jobs:
         run: yarn install
 
       - name: Builds Project ⚙️
-        run: yarn build
+        run: CONTENTFUL_SPACE_ID=${{ secrets.CONTENTFUL_SPACE_ID }} CONTENTFUL_ACCESS_TOKEN=${{ secrets.CONTENTFUL_ACCESS_TOKEN }} yarn build
 
       - name: Run Tests 🔬
         run: yarn test
