+++ core/api/contentful.ts
@@ -1,6 +1,7 @@
-import { createClient } from "contentful";
+import { createClient } from 'contentful';
 
+console.log(process.env.CONTENTFUL_SPACE_ID);
 export default createClient({
-  space: process.env.CONTENTFUL_SPACE_ID || "test",
-  accessToken: process.env.CONTENTFUL_ACCESS_TOKEN || "test",
+  space: process.env.CONTENTFUL_SPACE_ID || 'test',
+  accessToken: process.env.CONTENTFUL_ACCESS_TOKEN || 'test',
 });
