+++ package.json
@@ -44,6 +44,7 @@
     "@types/node": "^14.14.27",
     "@types/react": "^17.0.2",
     "babel-jest": "^26.6.3",
+    "cross-env": "^7.0.3",
     "gh-pages": "^3.1.0",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^26.6.3",
