+++ pages/contact.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React from 'react';
 
 const Contact = () => {
   return <div>Contact</div>;
