+++ pages/experience.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React from 'react';
 
 const Experience = () => {
   return <div>Experience</div>;
